requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
mysql-connector-python==9.0.0

# Data Validation
pydantic[email]==2.9.2
pydantic-settings==2.4.0

# Security
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9

# Development
python-dotenv==1.0.1

# Excel / Data Processing
numpy==2.1.1
openpyxl==3.1.5
pandas==2.2.3

# API Calls
requests==2.32.3